Examples using basic models

BreakingPendulum

BreakingPendulumInBox

DcMotorWithShaft

HalfWaveRectifier

StructuralHalfWaveRectifier

InitialCondition

MkinInitialCondition

Vanderpol

VanderpolWithEvents

VanderpolWithParameter

Examples using basic models

These are available in Sims.Examples.Basics.

Here is an example of use:

using Sims
m = Sims.Examples.Basics.Vanderpol()
v = sim(m, 50.0)

plot(v)

BreakingPendulum

Sims.Examples.Basics.BreakingPendulum — Function.

Models a pendulum that breaks at 5 secs. This model uses a StructuralEvent to switch between Pendulum mode and FreeFall mode.

Based on an example by George Giorgidze's thesis](http://eprints.nottingham.ac.uk/12554/1/main.pdf) that's in Hydra.

source

BreakingPendulumInBox

Sims.Examples.Basics.BreakingPendulumInBox — Function.

An extension of Sims.Examples.Basics.BreakingPendulum.

Floors and a wall are added. These are handled by Events in the FreeFall model. Velocities are reversed to bounce the ball.

source

DcMotorWithShaft

Sims.Examples.Basics.DcMotorWithShaft — Function.

A DC motor with a flexible shaft. The shaft is made of multiple elements. These are collected together algorithmically.

This is a smaller version of an example on p. 117 of David Broman's thesis.

I don't know if the results are reasonable or not.

source

HalfWaveRectifier

Sims.Examples.Basics.HalfWaveRectifier — Function.

A half-wave rectifier. The diode uses Events to toggle switching.

See F. E. Cellier and E. Kofman, Continuous System Simulation, Springer, 2006, fig 9.27.

source

StructuralHalfWaveRectifier

Sims.Examples.Basics.StructuralHalfWaveRectifier — Function.

This is the same circuit used in Sims.Examples.Basics.HalfWaveRectifier, but a structurally variable diode is used instead of a diode that uses Events.

source

InitialCondition

Sims.Examples.Basics.InitialCondition — Function.

A basic test of solving for initial conditions for two simultaineous equations.

source

MkinInitialCondition

Sims.Examples.Basics.MkinInitialCondition — Function.

A basic test of solving for initial conditions for two simultaineous equations.

source

Vanderpol

Sims.Examples.Basics.Vanderpol — Function.

The Van Der Pol oscillator is a simple problem with two equations and two unknowns.

source

VanderpolWithEvents

Sims.Examples.Basics.VanderpolWithEvents — Function.

An extension of Sims.Examples.Basics.Vanderpol. Events are triggered every 2 sec that change the quantity mu.

source

VanderpolWithParameter

Sims.Examples.Basics.VanderpolWithParameter — Function.

The Van Der Pol oscillator is a simple problem with two equations and two unknowns.

source