Examples using basic models

DcMotorWithShaft

Vanderpol

Examples using basic models

These are available in FunctionalModels.Examples.Basics.

Here is an example of use:

using FunctionalModels
m = FunctionalModels.Examples.Basics.Vanderpol()
v = sim(m, 50.0)

plot(v)

DcMotorWithShaft

Main.Examples.Basics.DcMotorWithShaft — Function

A DC motor with a flexible shaft. The shaft is made of multiple elements. These are collected together algorithmically.

This is a smaller version of an example on p. 117 of David Broman's thesis.

I don't know if the results are reasonable or not.

Vanderpol

Main.Examples.Basics.Vanderpol — Function

The Van Der Pol oscillator is a simple problem with two equations and two unknowns.