The Sims standard library

Signal

ElectricalNode

HeatPort

Flange

Current

Voltage

HeatFlow

Temperature

Torque

Angle

AngularVelocity

AngularAcceleration

The Sims standard library

NOTE: Many of these components are unfinished or broken. These are mainly components that require events and support of discrete systems.

These components are available with Sims.Lib.

Normal usage is:

using Sims
using Sims.Lib

# modeling...

Library components include models for:

Electrical circuits

Heat transfer

Rotational mechanics

Most of the components mimic those in the Modelica Standard Library.

The main types for Unknowns and signals defined in Sims.Lib include:

Flow/through variable	Node/across variable	Node helper type

Electrical systems	Current	Voltage	ElectricalNode

Heat transfer	HeatFlow	Temperature	HeatPort

Mechanical rotation	Torque	Angle	Flange

ModelingToolkit does not have variable types, so Signal, ElectricalNode, HeatPort, and Flange are all aliases of Any, mainly to help with documentation of models.

Current, Voltage, HeatFlow, Temperature, Torque, and Angle are all helper functions that create variables with gensym.

Signal

Sims.Lib.Signal — Type

Signal is an alias for Any used to indicate a signal value or variable.

ElectricalNode

Sims.Lib.ElectricalNode — Type

ElectricalNode is an alias for Any used to indicate a node voltage value or variable.

HeatPort

Sims.Lib.HeatPort — Type

HeatPort is an alias for Any used to indicate a value or variable for a temperature port.

Flange

Sims.Lib.Flange — Type

Flange is an alias for Any used to indicate a value or variable for a flange port (an angle).

Current

Sims.Lib.Current — Function

Current(x = 0.0; name = :i) creates an Unknown with a default name of :i.

Voltage

Sims.Lib.Voltage — Function

Voltage(x = 0.0; name = :i) creates an Unknown with a default name of :v.

HeatFlow

Sims.Lib.HeatFlow — Function

HeatFlow(x = 0.0; name = :i) creates an Unknown with a default name of :hf.

Temperature

Sims.Lib.Temperature — Function

Temperature(x = 0.0; name = :i) creates an Unknown with a default name of :T.

Torque

Sims.Lib.Torque — Function

Torque(x = 0.0; name = :i) creates an Unknown with a default name of :torque.

Angle

Sims.Lib.Angle — Function

Angle(x = 0.0; name = :i) creates an Unknown with a default name of :angle.

AngularVelocity

Sims.Lib.AngularVelocity — Function

AngularVelocity(x = 0.0; name = :i) creates an Unknown with a default name of :angvelocity.

AngularAcceleration

Sims.Lib.AngularAcceleration — Function

AngularAcceleration(x = 0.0; name = :i) creates an Unknown with a default name of :angacceleration.